made seed an integer

+++ src/simulator.py
@@ -236,7 +236,7 @@ class Simulator:
         if config is None:
             config = MapConfig()
         if seed is None:
-            seed = random.random() * 1_000_000_000
+            seed = (int) (random.random() * 1_000_000_000)
         print(f"Using seed {seed} to generate the map.")
         random.seed(seed)
         # generates and sets the 2D grid with agent and victims, currently with preset values.
